快照背景添加颜色和图片设置

+++ App预览图/Module/Screenshot/Model/ColorInfo.h
@@ -15,7 +15,7 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, assign) NSInteger red;
 @property (nonatomic, assign) NSInteger green;
 @property (nonatomic, assign) NSInteger blue;
-@property (nonatomic, assign) NSInteger opacity;
+@property (nonatomic, assign) CGFloat opacity;
 
 @end
